A cast is needed here to avoid warnings in both 32 bit and 64 bit
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@9562 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/surf/surf_action.c
@@ -68,7 +68,7 @@ void *surf_action_new(size_t size, double cost, surf_model_t model,
 {
   xbt_assert2(size <= action_mallocator_allocated_size,
       "Cannot create a surf action of size %lu: the mallocator only provides actions of size %d",
-      size, action_mallocator_allocated_size);
+      (long unsigned) size, action_mallocator_allocated_size);
 
   surf_action_t action = xbt_mallocator_get(action_mallocator);
   action->refcount = 1;
